+++ src/components/Header/Header.module.css
@@ -18,12 +18,24 @@
     white-space: nowrap;
 }
 
+.btn:hover {
+    opacity: 0.5;
+}
+
+.btn_outline:hover {
+    opacity: 1;
+}
+
 .user {
     margin: 0;
-    font-size: 24px;
-    line-height: 24px;
+    font-size: 14px;
+    font-weight: 500;
 }
 
 .light {
+    color: aliceblue;
+}
+
+.light:hover {
     color: aliceblue;
 }
