+++ src/index.tsx
@@ -5,7 +5,7 @@ import './index.css';
 import {Provider} from 'react-redux';
 import {App} from './components/App/App';
 import {store} from './store';
-import {ThemeProvider} from './services/ThemeProvider';
+import {ThemeProvider} from './services/theme/ThemeProvider';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
