+++ src/services/theme/ThemeClasses.ts
@@ -1,0 +1,4 @@
+export const ThemeClass = {
+    dark: 1,
+    light: 2,
+};
