+++ src/services/theme/ThemeProvider.tsx
@@ -1,4 +1,4 @@
-import React, {createContext, useMemo, useState} from 'react';
+import React, {createContext, useCallback, useMemo, useState} from 'react';
 
 type ThemeProps = {
     children: React.ReactNode;
@@ -11,12 +11,9 @@ export const ThemeContext = createContext({
 
 function ThemeProvider({children}: ThemeProps) {
     const [theme, setTheme] = useState('dark');
-    const toggleTheme = useMemo(
-        () => () => {
-            setTheme(theme === 'dark' ? 'light' : 'dark');
-        },
-        [theme]
-    );
+    const toggleTheme = useCallback(() => {
+        setTheme(theme === 'dark' ? 'light' : 'dark');
+    }, [theme]);
 
     const themeObject = useMemo(() => ({theme, toggleTheme}), [theme, toggleTheme]);
